refactor(session): isolate provider usage normalization
